disabling automatic quire validation to RCA the segfaults

+++ tests/quire_test_helpers.hpp
@@ -103,13 +103,13 @@ namespace sw {
 			quire<nbits, es, capacity> q;
 
 			// report some parameters about the posit and quire configuration
-			int max_scale = q.max_scale();
-			int min_scale = q.min_scale();
+			size_t max_scale = q.max_scale();
+			size_t min_scale = q.min_scale();
 			std::cout << "Maximum scale  = " << max_scale << " Minimum scale  = " << min_scale << " Dynamic range = " << q.dynamic_range() << std::endl;
 			std::cout << "Maxpos Squared = " << maxpos_scale<nbits, es>() * 2 << " Minpos Squared = " << minpos_scale<nbits, es>() * 2 << std::endl;
 
 			// cover the scales with one order outside of the dynamic range of the quire configuration (minpos^2 and maxpos^2)
-			for (int scale = max_scale + 1; scale >= min_scale - 1; scale--) {  // extend by 1 max and min scale to test edge of the quire
+			for (int scale = int(max_scale + 1); scale >= min_scale - 1; scale--) {  // extend by 1 max and min scale to test edge of the quire
 				value<fbits> v = std::pow(2.0, scale);
 				try {
 					q = v;
